app/src/main/java/com/girmiti/mobilepos/net/model/LoginResponse.java
package com.girmiti.mobilepos.net.model;

public class LoginResponse extends Response {

    private String terminalID;
    private String bussinessName;
    private String email;
    private CurrencyDTO currencyDTO;
    private TSMResponse terminalData;

    private String address;

    private String city;

    private String state;

    private String country;

    private String pin;

    public String getAddress() {
        return address;
    }

    public void setAddress(String address) {
        this.address = address;
    }

    public String getCity() {
        return city;
    }

    public void setCity(String city) {
        this.city = city;
    }

    public String getState() {
        return state;
    }

    public void setState(String state) {
        this.state = state;
    }

    public String getCountry() {
        return country;
    }

    public void setCountry(String country) {
        this.country = country;
    }

    public String getPin() {
        return pin;
    }

    public void setPin(String pin) {
        this.pin = pin;
    }

    public LoginResponse(String response) {
        super(response);
    }

    public String getTerminalID() {
        return terminalID;
    }

    public void setTerminalID(String terminalID) {
        this.terminalID = terminalID;
    }

    public String getBussinessName() {
        return bussinessName;
    }

    public void setBussinessName(String bussinessName) {
        this.bussinessName = bussinessName;
    }

    public CurrencyDTO getCurrencyDTO() {
        return currencyDTO;
    }

    public void setCurrencyDTO(CurrencyDTO currencyDTO) {
        this.currencyDTO = currencyDTO;
    }

    public TSMResponse getTerminalData() {
        return terminalData;
    }

    public void setTerminalData(TSMResponse terminalData) {
        this.terminalData = terminalData;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }
}
